requirements.txt
fastapi
pydantic
pymongo
python-dotenv
bcrypt
uvicorn
